requirements.txt
kurigram
TgCrypto
redis[hiredis]>=5.0.0
apscheduler>=3.10.0
aiohttp>=3.9.0
google-generativeai>=0.5.0
sqlalchemy>=2.0.0
pyyaml>=6.0
pytz
